requirements.txt
Django==3.2.10
django-bootstrap4==21.1
django-filter==21.1
djangorestframework==3.13.1
psycopg2==2.9.2
sqlparse==0.4.2
